Fix device detection (port scanning)
See https://github.com/micro-manager/micro-manager/issues/625. This is
an attempt to gracefully exit when port properties can't be read.

Currently failure happens when the port is a TCP/IP port because this
port doesn't have such properties and the whole detection process is
aborted (even if the device has been found).

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@16706 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd