Core: No longer throws an uncaught exception when calling getSLMWidth/Height with a non-existing SLM.
Previously, such a call would result in MM disappearing from the screen.  The two functions are now 
declared as throwing CMMError, which causes the Swig layer to throw and exception that can be 
caused in the Java layer.  

Since this changes the API (and its behavior), the Core version was increased to 9.0.0.

git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17080 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd