Deprecate MMCore property block functions
It is not clear to me what these were intended for, but they are
probably part of a planned but unfinished feature (seeing that it
appears to be tied to an Equipment command in config files, which, as
far as I know, was never supported by the Hardware Configuration
Wizard).

Some of the functions deprecated here may crash when called with
incorrect parameters.