Experimental Meson build for MMDevice, MMCore
MMDevice builds on its own.

MMCore requires a copy (or symlink) of MMDevice to be in its
subprojects/ directory. (This is unavoidable as we develop the Meson
build in place. When MMDevice and MMCore are in their own repositories
(and/or have source tarballs) MMCore will be able to depend on MMDevice
as either a git submodule or a Meson wrap.)