MMCore: List all transitively public headers
(Not used in any build yet.)

Currently MMCore is only used to build MMCoreJ and pymmcore, and
providing a good (minimal) set of public headers (as with proper support
to use MMCore as a C++ library by itself) should probably be a future
goal. For now, list the headers needed in practice.