Check for null serialized metadata from camera
In #710 (32057fdc9b2b479c4a41e7ca229f6b470f546520) I changed the
signature of one of the two overloads of InsertImage() from

virtual int InsertImage(
    const Device* caller,
    const unsigned char* buf,
    unsigned width,
    unsigned height,
    unsigned byteDepth,
    const char* serializedMetadata,
    const bool doProcess = true) = 0;

to

virtual int InsertImage(
    const Device* caller,
    const unsigned char* buf,
    unsigned width,
    unsigned height,
    unsigned byteDepth,
    const char* serializedMetadata = nullptr,
    const bool doProcess = true) = 0;

(added nullptr default argument to serializedMetadata).

This was to support existing cameras that were calling the deleted
overload

MM_DEPRECATED(
virtual int InsertImage(
    const Device* caller,
    const unsigned char* buf,
    unsigned width,
    unsigned height,
    unsigned byteDepth,
    const Metadata* md = 0,
    const bool doProcess = true)) = 0;

without passing a non-null `md`.

But I forgot to ensure that the called implementations in CoreCallback
is checking for nullptr in this case (actually forgot to include the
relevant commit). So here is a fix.