Meson: MMCore: default to hidden visibility
This does to MMCore what #735 did to MMDevice.

We currently build MMCore only as a static library (never shared
library/DLL on its own), but the primary use case is to build it into a
shared library (Python or Java extension). It is almost never the case
that a shared library wants to expose the MMCore C++ API to its clients.
So this is a good default.

If we (in the future) add support for building MMCore as a C++ shared
library, this is still a good setting, because then each public API
function should be explicitly exported (as it will have to be on Windows
anyway).